+++ src/modules/preprocessor/__tests__/preprocessor.ts
@@ -1,10 +1,10 @@
 import es from 'estree'
 
-import { parseError } from '../../index'
-import { mockContext } from '../../mocks/context'
-import { parse } from '../../parser/parser'
-import { accessExportFunctionName, defaultExportLookupName } from '../../stdlib/localImport.prelude'
-import { Chapter } from '../../types'
+import { parseError } from '../../../index'
+import { mockContext } from '../../../mocks/context'
+import { parse } from '../../../parser/parser'
+import { accessExportFunctionName, defaultExportLookupName } from '../../../stdlib/localImport.prelude'
+import { Chapter } from '../../../types'
 import preprocessFileImports, { getImportedLocalModulePaths } from '../preprocessor'
 import { parseCodeError, stripLocationInfo } from './utils'
