+++ src/modules/preprocessor/__tests__/transformers/hoistAndMergeImports.ts
@@ -1,6 +1,6 @@
-import { mockContext } from '../../../mocks/context'
-import { parse } from '../../../parser/parser'
-import { Chapter } from '../../../types'
+import { mockContext } from '../../../../mocks/context'
+import { parse } from '../../../../parser/parser'
+import { Chapter } from '../../../../types'
 import { hoistAndMergeImports } from '../../transformers/hoistAndMergeImports'
 import { parseCodeError, stripLocationInfo } from '../utils'
