+++ src/modules/preprocessor/__tests__/transformers/transformProgramToFunctionDeclaration.ts
@@ -1,7 +1,7 @@
-import { mockContext } from '../../../mocks/context'
-import { parse } from '../../../parser/parser'
-import { defaultExportLookupName } from '../../../stdlib/localImport.prelude'
-import { Chapter } from '../../../types'
+import { mockContext } from '../../../../mocks/context'
+import { parse } from '../../../../parser/parser'
+import { defaultExportLookupName } from '../../../../stdlib/localImport.prelude'
+import { Chapter } from '../../../../types'
 import { transformProgramToFunctionDeclaration } from '../../transformers/transformProgramToFunctionDeclaration'
 import { parseCodeError, stripLocationInfo } from '../utils'
