+++ src/modules/preprocessor/__tests__/utils.ts
@@ -1,6 +1,6 @@
 import es from 'estree'
 
-import { full, simple } from '../../utils/walkers'
+import { full, simple } from '../../../utils/walkers'
 
 export const parseCodeError = new Error('Unable to parse code')
