+++ src/modules/preprocessor/transformers/hoistAndMergeImports.ts
