+++ src/modules/preprocessor/transformers/removeNonSourceModuleImports.ts
@@ -1,6 +1,6 @@
 import es from 'estree'
 
-import { ancestor } from '../../utils/walkers'
+import { ancestor } from '../../../utils/walkers'
 import { isFilePath } from '../filePaths'
 
 /**
