+++ src/modules/preprocessor/typeGuards.ts
